move api-keys to server. confirmed working
